refactor: 사용자의 요구사항 중 추상적인 날짜 제외 및 편의시설 제외한 모든 요청 가능

+++ src/main/java/team/mjk/agent/domain/agoda/dto/request/Additional.java
@@ -5,15 +5,16 @@ import lombok.Builder;
 @Builder
 public record Additional(
 
-        String currency,
-        DailyRate dailyRate,
-        boolean discountOnly,
-        String language,
-        int maxResult,
-        int minimumReviewScore,
-        int minimumStarRating,
-        Occupancy occupancy,
-        String sortBy
+    String currency,
+    DailyRate dailyRate,
+    boolean discountOnly,
+    String language,
+    int maxResult,
+    double minimumReviewScore,
+    double minimumStarRating,
+    Occupancy occupancy,
+    String sortBy
 
 ) {
+
 }
